# challenges/challenge-1-web-api/python-fastapi/openapi.yaml
openapi: "3.0.3"
info:
  title: EDP Field Ops — Task Manager API
  version: "1.0.0"
  description: |
    REST API for managing field operations work orders at EDP.
    Track outages, grid maintenance, solar installations, EV charging deployments and safety inspections.
servers:
  - url: http://127.0.0.1:8000
    description: Local development server

# ── Security ───────────────────────────────────────────────────────────────
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable schemas ────────────────────────────────────────────────────
  schemas:
...
    UserRole:
      type: string
      enum: [viewer, user, admin]
...
    UserRegister:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
    UserLogin:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
          maxLength: 128
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          default: bearer

    # Tasks
    CategoryResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
...
    CategoryCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    BulkCategorizeRequest:
      type: object
      required: [task_ids, category_id]
      properties:
        task_ids:
          type: array
          items:
            type: integer
          minItems: 1
        category_id:
          type: integer
    BulkCategorizeResponse:
      type: object
      properties:
        updated:
          type: integer
        category:
          $ref: "#/components/schemas/CategoryResponse"

    # Comments
...
    HTTPError:
      type: object
      properties:
        detail:
          type: string
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  oneOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string

  # ── Reusable responses ──────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
          example:
            detail: "Not authenticated"
    Forbidden:
      description: Authenticated but insufficient role or not the owner
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
          example:
            detail: "You do not have permission to modify this task"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
          example:
            detail: "Task not found"
    UnprocessableEntity:
      description: Request body failed Pydantic validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationError"
    Conflict:
      description: Unique constraint violation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
    RateLimited:
      description: Rate limit exceeded (slowapi)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/HTTPError"
          example:
            detail: "Rate limit exceeded: 5 per 1 minute"

# ── Global security (overridden per-route where not required) ───────────────
security:
  - BearerAuth: []

# ============================================================================
# Paths
# ============================================================================
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      description: |
        Creates a new account. The assigned role is always `viewer`.
        Rate-limited to **5 requests / minute** per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
            example:
              email: technician@edp.pt
              password: securepass123
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              example:
                detail: Email already registered
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/login:
    post:
      tags: [auth]
      summary: Login and obtain a JWT token
      description: |
        Authenticates with email + password and returns a signed JWT (HS256, 60 min).
        The token embeds `user_id` and `role`.
        Rate-limited to **10 requests / minute** per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
            example:
              email: technician@edp.pt
              password: securepass123
      responses:
        "200":
          description: JWT token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
              example:
                detail: Invalid email or password
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/RateLimited"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh JWT token
      description: |
        Issues a new token from a valid existing token.
        The role is always re-read from the database, so a demoted user immediately loses elevated access.
      responses:
        "200":
          description: New JWT token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Tasks ─────────────────────────────────────────────────────────────────
...
  /tasks/bulk/categorize:
    post:
      tags: [categories]
      summary: Bulk assign a category to multiple tasks
      description: |
        Atomically assigns a category to all specified tasks.
        If ANY task ID is not found or the user is not the owner, the entire operation rolls back.
        Requires `user` or `admin` role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCategorizeRequest"
            example:
              task_ids: [1, 2, 3]
              category_id: 5
      responses:
        "200":
          description: Tasks categorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkCategorizeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /categories:
    get:
      tags: [categories]
      summary: List all categories
      description: Returns all category labels ordered alphabetically. All authenticated roles.
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [categories]
      summary: Create a category
      description: Requires `admin` role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreate"
            example:
              name: grid_maintenance
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

  # ── Admin ─────────────────────────────────────────────────────────────────
...